Move python assignments into pre-work repo
Previously the intermediate python assignments were all in separate repos. As this pre-work repo is currently public, for easier distribution to students, and the assignment repos were private, simply putting all the assignments in the pre-work repo makes more sense.

Later when our overall github organization is more settled & we can be confident assignments will always live in the same github org, we can see if we want to have these link out to assignment repos for easier updating, etc. Not a pressing concern right now.